[META] Port report menu additions to 10
Also I have restructured the module to have a normal structure.
The report definitions for the 3 reports which were in version 8 are missing
Also, there are reports which will be added to this repo which I have not yet
created skeletons for.
Just milestoning progress at this point
