+++ UK_Reports/models/account_invoice.py
@@ -3,7 +3,7 @@
 ##############################################################################
 #
 # UK Report Template
-# Copyright (C) 2015 OpusVL (<http://opusvl.com/>)
+# Copyright (C) 2019 OpusVL (<http://opusvl.com/>)
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
@@ -20,33 +20,20 @@
 #
 ##############################################################################
 
+from odoo import models, api
 
 
-{
-   'name': 'UK customised reports',
-   'version': '0.1',
-   'category': 'Reporting',
-   'description': """
-   Customised reports for UK:
-    - Sales order / Quotation
-    - Account balance / overdue
-    - Invoice print 
+class AccountInvoice(models.Model):
+	_inherit = "account.invoice"
 
-    These have been migrated from V3.3, V4, V5, V6, V6.1 to V8
+	@api.multi
+	def invoice_print(self):
+		""" Print the invoice and mark it as sent, so that we can see more
+			easily the next step of the workflow
+		"""
+		self.sent = True
+		self.ensure_one()
+		return self.env['report'].get_action(self, 'account.UK_Invoice')
 
-   """,
-   'author': 'OpusVL',
-   'website': 'http://opusvl.com',
-   'depends': ['account_accountant', 'sale', 'report'],
-   'init_xml': [],
-   'update_xml': [
-       'reports.xml',
-   ],
-   'demo_xml': [],
-   'test': [],
-   'license': 'AGPL-3',
-   'installable': True,
-   'active': False,
-}
 
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
