--- UK_Reports/reports.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <report auto="False" id="sale.order_print_UK" model="sale.order" name="sale.UK_SalesOrder" rml="UK_Reports/report/UK_sale_print_order.rml" string="Sales Order (UK format)" />
-        <report auto="False" id="account.invoice_print_UK" model="account.invoice" name="account.UK_Invoice" rml="UK_Reports/report/UK_account_print_invoice.rml" string="Invoice (UK format)" />
-        <report auto="False" id="account.statement_print_UK" model="res.partner" name="account.UK_Statement.2" rml="UK_Reports/report/UK_account_print_overdue.rml" string="Account statement (UK format)" />
-    </data>
-</openerp>
-
